Found out that a lot of GoL people use a specific format for their stuff.
So I added in a section to split on how to load the file based on the file
extenstion. If it ends with gol then its the original straight 01 for bit state
implemenetation. If it ends in rle it is in the run length encoding format that
I have found lots of files and patterns for.