Minor fixes and additions
Needed to add in a check for empty list of files, and exit with message if non
given.

Added in help text to existing cli flags. I also added in a flag for the format
to take the file in, and it'll rely on this if file doesn't have a proper
extension.